refactor(scheduler): split env parsing helpers
