+++ sglang_omni/scheduling/observability.py
@@ -3,28 +3,7 @@
 
 from __future__ import annotations
 
-import os
 from typing import Any
-
-
-def env_float(name: str, *, default: float) -> float:
-    value = os.environ.get(name)
-    if value is None or value.strip() == "":
-        return default
-    try:
-        return float(value)
-    except ValueError:
-        return default
-
-
-def env_int(name: str) -> int | None:
-    value = os.environ.get(name)
-    if value is None or value.strip() == "":
-        return None
-    try:
-        return int(value)
-    except ValueError:
-        return None
 
 
 def safe_qsize(queue_obj: Any) -> int | None:
